# back-end/public/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "아팠찌 페이지 json 형식",
        "description": "이 프로젝트는 Laravel 백엔드 기반 API 문서입니다.",
        "version": "1.0.0"
    },
    "paths": {
        "/api/ai/code-review": {
            "post": {
                "tags": [
                    "AI"
                ],
                "summary": "AI 코드 리뷰 요청",
                "description": "입력된 언어, 코드, 상황을 기반으로 GPT 모델을 이용해 코드 리뷰를 수행합니다.",
                "operationId": "aiCodeReview",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "language",
                                    "code",
                                    "situation"
                                ],
                                "properties": {
                                    "language": {
                                        "type": "string",
                                        "example": "java"
                                    },
                                    "code": {
                                        "type": "string",
                                        "example": "int a = 3; int b = 10; printf(\\'%%d\\', a/b);"
                                    },
                                    "situation": {
                                        "type": "string",
                                        "example": "실수가 아니라 정수가 나옴"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "AI 코드 리뷰 성공",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "result": {
                                            "properties": {
                                                "id": {
                                                    "type": "string",
                                                    "example": "chatcmpl-BePb0p6PgWwvh9IRkGEUIN9K7UCR6"
                                                },
                                                "object": {
                                                    "type": "string",
                                                    "example": "chat.completion"
                                                },
                                                "created": {
                                                    "type": "integer",
                                                    "example": 1748971210
                                                },
                                                "model": {
                                                    "type": "string",
                                                    "example": "gpt-4o-mini-2024-07-18"
                                                },
                                                "choices": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "index": {
                                                                "type": "integer",
                                                                "example": 0
                                                            },
                                                            "message": {
                                                                "properties": {
                                                                    "role": {
                                                                        "type": "string",
                                                                        "example": "assistant"
                                                                    },
                                                                    "content": {
                                                                        "type": "string",
                                                                        "example": "Java에서는 두 개의 정수(int) 타입을 나누면 결과도 정수형으로..."
                                                                    }
                                                                },
                                                                "type": "object"
                                                            },
                                                            "finish_reason": {
                                                                "type": "string",
                                                                "example": "stop"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "usage": {
                                                    "properties": {
                                                        "prompt_tokens": {
                                                            "type": "integer",
                                                            "example": 78
                                                        },
                                                        "completion_tokens": {
                                                            "type": "integer",
                                                            "example": 340
                                                        },
                                                        "total_tokens": {
                                                            "type": "integer",
                                                            "example": 418
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "API 호출 실패",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Moderation API 호출 실패"
                                        },
                                        "debug": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}